Rename 'applications' to 'subcommands'
