+++ hod/rmscheduler/job.py
@@ -61,8 +61,6 @@ class Job(object):
     def generate_script(self):
         """Build the submit script"""
         script = ["#!/bin/bash", "## Generated by HOD"]
-        print 'script', script
-        print 'self', self
         script += self.generate_script_header()
 
         script += self.generate_environment()
