+++ test/unit/subcommands/test_applications_pbs.py
@@ -28,16 +28,37 @@
 """
 
 import unittest
-import hod.applications.genconfig as hag
+import pytest
 
-from mock import Mock
+from hod.subcommands.pbs import CreatePbsApplication
 
-class TestGenconfigApplication(unittest.TestCase):
-    def test_run(self):
-        app = hag.GenConfigApplication()
-        app.run([])
+from mock import patch
+
+class TestCreatePbsApplication(unittest.TestCase):
+    @pytest.mark.xfail(reason="Requires easybuild environment")
+    def test_run_no_args(self):
+        with patch('hod.subcommands.pbs.PbsHodJob'):
+            app = CreatePbsApplication()
+            self.assertRaises(ValueError, app.run, [])
+
+    @pytest.mark.xfail(reason="Requires easybuild environment")
+    def test_run_with_args(self):
+        with patch('hod.subcommands.pbs.PbsHodJob'):
+            app = CreatePbsApplication()
+            app.run(['--config=hod.conf', '--workdir=workdir'])
+
+    def test_run_with_dist_arg(self):
+        with patch('hod.subcommands.pbs.PbsHodJob'):
+            app = CreatePbsApplication()
+            app.run(['--dist=Hadoop-2.3.0', '--workdir=workdir'])
+
+    def test_run_fails_with_config_and_dist_arg(self):
+        with patch('hod.subcommands.pbs.PbsHodJob'):
+            app = CreatePbsApplication()
+            self.assertRaises(ValueError, app.run, 
+                    ['--config=hod.conf', '--dist=Hadoop-2.3.0', '--workdir=workdir'])
 
     def test_usage(self):
-        app = hag.GenConfigApplication()
+        app = CreatePbsApplication()
         usage = app.usage()
         self.assertTrue(isinstance(usage, basestring))
