--- test/unit/test_applications_genconfig.py
@@ -1,5 +1,0 @@
-import unittest
-import hod.applications.genconfig as hag
-
-class TestGenconfigApplication(unittest.TestCase):
-    pass
